+++ test/unit_tests/test_stack.py
@@ -3,7 +3,7 @@ from nose.tools import *
 from troposphere import Tags, Ref
 
 userdata = keypair = instance_type = code_deploy_service_role = vpc_cidr = public_cidr = \
-    minsize = maxsize = path2ping = nat_image_id = jump_image_id = unit_image_id = health_check_grace_period = \
+    minsize = maxsize = elb_health_check = nat_image_id = jump_image_id = unit_image_id = health_check_grace_period = \
     health_check_type = db_instance_type = db_engine = db_port = db_hdd_size = owner_emails = nat_alerting = \
     db_backup_window = db_backup_retention = db_maintenance_window = db_storage_type = block_devices_config = None
 availability_zones = []
@@ -15,7 +15,7 @@ protocols = []
 
 def setup_resources():
     global userdata, availability_zones, keypair, instance_type, code_deploy_service_role, vpc_cidr, \
-        public_cidr, instanceports, loadbalancerports, protocols, minsize, maxsize, path2ping, home_cidrs, \
+        public_cidr, instanceports, loadbalancerports, protocols, minsize, maxsize, elb_health_check, home_cidrs, \
         nat_image_id, jump_image_id, health_check_grace_period, health_check_type, unit_image_id, db_instance_type, \
         db_engine, db_port, db_hdd_size, owner_emails, nat_alerting, db_backup_window, db_backup_retention, \
         db_maintenance_window, db_storage_type, block_devices_config
@@ -43,7 +43,7 @@ runcmd:
     protocols = ['HTTP']
     minsize = 1
     maxsize = 1
-    path2ping = '/index.html'
+    elb_health_check = 'HTTP:80/index.html'
     public_cidr = {'name': 'PublicIp', 'cidr': '0.0.0.0/0'}
     health_check_grace_period = 300
     health_check_type = 'ELB'
@@ -160,10 +160,11 @@ def test_duplicate_unit_names():
                                    'protocols': protocols,
                                    'instanceports': instanceports,
                                    'loadbalancerports': loadbalancerports,
-                                   'path2ping': path2ping,
+                                   'elb_health_check': elb_health_check,
                                    'elb_log_bucket': None,
                                    'public_unit': True,
-                                   'unit_hosted_zone_name': None
+                                   'unit_hosted_zone_name': None,
+                                   'ssl_certificate_id': None
                                },
                                'dependencies': ['app2', 'db1'],
                                },
@@ -172,10 +173,11 @@ def test_duplicate_unit_names():
                                    'protocols': protocols,
                                    'instanceports': instanceports,
                                    'loadbalancerports': loadbalancerports,
-                                   'path2ping': path2ping,
+                                   'elb_health_check': elb_health_check,
                                    'elb_log_bucket': None,
                                    'public_unit': True,
-                                   'unit_hosted_zone_name': None
+                                   'unit_hosted_zone_name': None,
+                                   'ssl_certificate_id': None
                                },
                                'asg_config': {
                                    'minsize': minsize,
@@ -266,10 +268,11 @@ def test_duplicate_unit_names():
                                    'protocols': protocols,
                                    'instanceports': instanceports,
                                    'loadbalancerports': loadbalancerports,
-                                   'path2ping': path2ping,
+                                   'elb_health_check': elb_health_check,
                                    'unit_hosted_zone_name': None,
                                    'elb_log_bucket': None,
                                    'public_unit': True,
+                                   'ssl_certificate_id': None
                                },
                                'asg_config': {
                                    'minsize': minsize,
@@ -323,10 +326,11 @@ def test_duplicate_unit_names():
                                       'protocols': protocols,
                                       'instanceports': instanceports,
                                       'loadbalancerports': loadbalancerports,
-                                      'path2ping': path2ping,
+                                      'elb_health_check': elb_health_check,
                                       'unit_hosted_zone_name': None,
                                       'elb_log_bucket': None,
                                       'public_unit': True,
+                                      'ssl_certificate_id': None
                                   },
                                   'blue_asg_config': {
                                       'minsize': minsize,
@@ -362,10 +366,11 @@ def test_duplicate_unit_names():
                                       'protocols': protocols,
                                       'instanceports': instanceports,
                                       'loadbalancerports': loadbalancerports,
-                                      'path2ping': path2ping,
+                                      'elb_health_check': elb_health_check,
                                       'unit_hosted_zone_name': None,
                                       'elb_log_bucket': None,
                                       'public_unit': True,
+                                      'ssl_certificate_id': None
                                   },
                                   'blue_asg_config': {
                                       'minsize': minsize,
@@ -409,7 +414,7 @@ def create_stack(stack_title):
     :return new stack
     """
     global userdata, availability_zones, keypair, instance_type, code_deploy_service_role, vpc_cidr, \
-        public_cidr, instanceports, loadbalancerports, protocols, minsize, maxsize, path2ping, home_cidrs, \
+        public_cidr, instanceports, loadbalancerports, protocols, minsize, maxsize, elb_health_check, home_cidrs, \
         nat_image_id, jump_image_id, health_check_grace_period, health_check_type, unit_image_id, db_instance_type, \
         db_engine, db_port, owner_emails, nat_alerting, db_backup_window, db_backup_retention, db_maintenance_window, \
         db_storage_type, block_devices_config
@@ -435,10 +440,11 @@ def create_stack(stack_title):
                                    'protocols': protocols,
                                    'instanceports': instanceports,
                                    'loadbalancerports': loadbalancerports,
-                                   'path2ping': path2ping,
+                                   'elb_health_check': elb_health_check,
                                    'unit_hosted_zone_name': None,
                                    'elb_log_bucket': None,
                                    'public_unit': True,
+                                   'ssl_certificate_id': None
                                },
                                'blue_asg_config': {
                                    'minsize': minsize,
@@ -474,10 +480,11 @@ def create_stack(stack_title):
                                 'protocols': protocols,
                                 'instanceports': instanceports,
                                 'loadbalancerports': loadbalancerports,
-                                'path2ping': path2ping,
+                                'elb_health_check': elb_health_check,
                                 'unit_hosted_zone_name': None,
                                 'elb_log_bucket': None,
                                 'public_unit': True,
+                                'ssl_certificate_id': None
                             },
                             'asg_config': {
                                 'minsize': minsize,
@@ -499,10 +506,11 @@ def create_stack(stack_title):
                                 'protocols': protocols,
                                 'instanceports': instanceports,
                                 'loadbalancerports': loadbalancerports,
-                                'path2ping': path2ping,
+                                'elb_health_check': elb_health_check,
                                 'unit_hosted_zone_name': None,
                                 'elb_log_bucket': None,
                                 'public_unit': True,
+                                'ssl_certificate_id': None
                             },
                             'asg_config': {
                                 'minsize': minsize,